third_party/blink/web_tests/external/wpt/html/rendering/non-replaced-elements/form-controls/resets-expected.txt

This is a testharness.js-based test.
Found 3 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] <input type="image"> - overflow
  assert_equals: expected "clip" but got "visible"
[FAIL] <input type="image"> - overflow-clip-margin
  assert_equals: expected "content-box" but got "0px"
[FAIL] <textarea> - overflow
  assert_equals: expected "visible" but got "auto"
Harness: the test ran to completion.
